requirements.txt
figlet
termcolor
argparse
requests
